Feat: Handle inputs validation and Add toast in AddEditScreen.
AddEditViewModel - add _state variable.

REF: change values from Int to Double in Receipt and Product

changename from receiptsList to List

+++ app/src/main/java/com/cleverexpenses/receipts/feature_receipt/domain/model/Product.kt
@@ -14,8 +14,8 @@ import androidx.room.PrimaryKey
 data class Product(
     @PrimaryKey val productId: Int,
     val name: String,
-    val quantity: Int,
-    val unitPrice: Int,
-    val totalPrice: Int,
+    val quantity: Double,
+    val unitPrice: Double,
+    val totalPrice: Double,
     val receiptOwnerId: Int
 )
